+++ proxy/noquic.go
@@ -57,3 +57,7 @@ func (be *Backend) http3Transport() http.RoundTripper {
 func http3Server(http.Handler) io.Closer {
 	return nil
 }
+
+func quicEndCopy(out, in net.Conn, writeErr, readErr *error) (bool, bool) {
+	return false, false
+}