+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # TLSPROXY Release Notes
 
+## v0.17.0
+
+### :wrench: Misc
+
+* Enable QUIC Stream Resets with Partial Delivery.
+* Forward QUIC stream errors.
+* Update go dependencies:
+  * upgraded github.com/c2FmZQ/http3-go v0.53.0 => v0.54.0
+  * upgraded github.com/c2FmZQ/quic-api v0.53.0 => v0.54.0
+  * upgraded github.com/quic-go/quic-go v0.53.0 => v0.54.0
+
 ## v0.16.3
 
 ### :wrench: Misc